Move getAllRegisteredDNodes to typescript file
